Fixes OTR loading (tested in SoH)
Previously it was failing with a file not found error on SFileOpenArchive,
for some reason removing `UNICODE` and `_UNICODE` from the compile definitions
of libultraship fixes it. It also fixes the issue of OTRExporter outputting
an OTR with only the first character of the filename present. (Said OTR was
still valid though, despite the odd name. Fixed by this regardless).