Consolidates compiler and linker flags in cmake.
Several compiler options were duplicated across different systems
unnecessarily. For example Windows "x64" and "Win32" were very nearly
identical already, and the couple of differences didn't make much sense.
We were passing /std:c++latest, but only to x64 Release. I just removed
that one altogether since we already set `PROPERTY CXX_STANDARD 20`
at the top of the file. We were also passing `/W3` regardless of configuration
on "x64", but on "Win32" we were passing `/W2` to Debug and `/W3` to Release.
If there's a reason for that let me know and it can be re-introduced.