soh build fixes (#14)
* changes needed to run `ExtractAssets`

* soh build stuff

* Make some things public

* dcvz changes

* More fixes

* fix pulseaudio

* windows static lib stuff

* fix soh build on windows

* Don’t require Threads for Wii U

* Wii U specific changes to upstream StormLib

* Fix up storm linking

* Add Wii U additions to spdlog

* More Wii U updates to libultraship

* Wii U specific import fixes

* LUS Wii U fixes

* STBI_NO_THREAD_LOCALS for Wii U

* Disable threads for Wii U

* Fixes OTR loading (tested in SoH)

Previously it was failing with a file not found error on SFileOpenArchive,
for some reason removing `UNICODE` and `_UNICODE` from the compile definitions
of libultraship fixes it. It also fixes the issue of OTRExporter outputting
an OTR with only the first character of the filename present. (Said OTR was
still valid though, despite the odd name. Fixed by this regardless).

* Re-adds missing Debug/Release cmake config and fixes Typos

This re-enables debugging libultraship files on Windows. Also fixes some
typos that previously weren't being compiled due to the missing
Debug/Release configuration.

* Consolidates compiler and linker flags in cmake.

Several compiler options were duplicated across different systems
unnecessarily. For example Windows "x64" and "Win32" were very nearly
identical already, and the couple of differences didn't make much sense.
We were passing /std:c++latest, but only to x64 Release. I just removed
that one altogether since we already set `PROPERTY CXX_STANDARD 20`
at the top of the file. We were also passing `/W3` regardless of configuration
on "x64", but on "Win32" we were passing `/W2` to Debug and `/W3` to Release.
If there's a reason for that let me know and it can be re-introduced.

* Embed STB into LUS

* Fixes Switch builds. (#4)

* Restricts the stormlib compile definition change to switch (#5)

* make tidy happy?

* actually make tidy happy

* checkout v3?

* remove commented out lines from cmakelists

* fontData not font_data

* suppress error

* add a space

Co-authored-by: briaguya <briaguya@alice>
Co-authored-by: David Chavez <david@dcvz.io>
Co-authored-by: Christopher Leggett <chris@leggett.dev>
Co-authored-by: David Chavez <davi@dcvz.io>
Co-authored-by: briaguya <briaguya>