Minor fixes (#327)
* Do not use available interface version to check for Waitable SwapChain

It's entirely possible for the D3D runtime to implement IDXGISwapChain2 while the SwapChain was never created with the appropriate flag to use a frame latency waitable object.

ALWAYS check the SwapChain flags.

* Prevent deadlock during gfx_dxgi_set_cursor_visibility if no mouse is present

If you continue calling `ShowCursor (...)` until you "get the value you want" and there is no mouse attached, it will lock the software up because Windows always returns -1 if there is no mouse.

* Avoid recreating swap chain on jitter fix changes

* Prevent infinite waiting when reducing max latency

This couldn't ever happen by just switching between 2 and 1 maximum frame latency. Only if it was reduced twice, like setting it to 4, then 3 and then 2. This oddball behavior is undocumented...

* No magic numbers pls

---------

Co-authored-by: Andon M. Coleman <Kaldaien@users.noreply.github.com>