Display handling (#326)
* Get refresh rate from display

Fixes "Match Refresh Rate" button. Gets the refresh rate from the monitor at the top left corner of the window.
Thanks to Aemony, who showed me how to do that without completely destroying performance while moving the window.

Co-Authored-By: Aemony <Aemony@users.noreply.github.com>

* Fallback to primary monitor if out of bounds

* Use current display for fullscreen size (SDL)

Used to always use the dimensions of the primary monitor, which of course lead to weird behavior, when not on the primary monitor.

* Out of bounds check for window (SDL)

* Restore window position on init

This should start fullscreen on the last monitor it was when exiting. should be save now that out of bounds checks are in place.

* Additional OOB check, just to make sure

---------

Co-authored-by: Aemony <Aemony@users.noreply.github.com>