Window positon tweaks and fullscreen in OpenGL (#333)
* Fixed window pos not restored with OpenGL

* Fixed fullscreen in OpenGL (again)

Fixes going full screen on the next monitor when the game was on the second half of the screen.

* Set Fullscreen state only when call succeeds

Also some logging so we hopefully know why it failed.

* Consistent behavior for getting monitor

SDL always gets the monitor  at the middle of the window. Now DXGI does too. Also only resets the position if there is no pixel of the window on any screen.

* also get current monitor on resize (DirectX)

* Correct typo

Co-authored-by: Kenix3 <kenixwhisperwind@gmail.com>

---------

Co-authored-by: Kenix3 <kenixwhisperwind@gmail.com>