Major rework and simplification of the code.
Made new discoveries and ended up not needing the scaling fix at all as the issue lay in the use, or rather not the use of a framebuffer at 1:1 internal resolution and the viewport.  With that out of the way this should fix the remaining scaling issues on macOS and bring it to parity with Windows and Linux functionality wise.