Mac fullscreen fix (#817)
* Implementation of the macOS fullscreen fixes

* Comments and small updates to fullscreen fix

Fixes the issue where windowed fullscreen cursor hiding didn't work as it should according to the presence of the imGUI menubar.
We still need a specific fix in each game to disable changing fullscreen mode while in fullscreen, but this can be removed when SDL3 gets implemented in the future.

Also renamed some macOS specific functions and variables to make sure it is clear that they are macOS only.

* Fixed an oversight from a previous version

* tidy-format, yada yada yada...

* Feel free to shoot me. I'm a professional.

* Trying to appease one of tidy-formats complaints