Dxgi: Remove unnecessary flushes and cleanup (#813)
* Clean up after #325

This cleans up some stuff left by https://github.com/Kenix3/libultraship/pull/325.
Most of that probably got optimized out anyway.

* Remove unnecessary Flush

The order is
    mRapi->EndFrame();
    mWapi->SwapBuffersBegin(); <- we are here
    mRapi->FinishRender();
    mWapi->SwapBuffersEnd();
and  mRapi->EndFrame(); is already using Flush(); so no need to do it twice in a row, with nothing else in between happening.

* Remove other unnecessary Flush

The order is
    mRapi->EndFrame();
    mWapi->SwapBuffersBegin();
    mRapi->FinishRender(); <- we are here
    mWapi->SwapBuffersEnd();

swap_chain->Present(mVsyncEnabled, DXGI_PRESENT_ALLOW_TEARING) in mWapi->SwapBuffersBegin(); is already an implicit Flush(); so no need to do it twice in a row, with nothing else in between happening.

* No flush, no include

<d3d11.h> and namespace Microsoft::WRL are not needed anymore without the flush in gfx_dxgi.cpp